Create README.md
64-bit Windows desktop application that displays the ADA/USD exchange rate.

Viewing the exchange rate of several currencies simultaneously with the ADA/USD exchange rate often causes anxiety and prevents a sensible perception of the market. With this in mind, we have created a Windows desktop application that displays the ADA/USD exchange rate. This will help you learn about Cardano's performance in a more analytical and less emotional way.